# openapi.yaml
openapi: 3.0.3

info:
  title: Calendar
  version: 1.0.0
  description: REST API de lembretes
  
paths:
  /reminders:
    get:
      operationId: getReminders
      summary: Consulta lembretes
      tags: 
        - Reminders
      parameters:
        - schema:
            type: string
            example: 'Treino Jiu Jitsu'
          name: tile
          description: 'Titutlo'
          in: query
          required: false
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderSummary'
  
    post:
      operationId: createReminders
      summary: Cria um novo lembrete
      tags: 
        - Reminders
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderInput'
      responses:
        '201':
          description: Lembrete criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  
  /reminder/{reminderId}:
    parameters:
      - schema:
          type: integer
          format: int64
        name: reminderId
        in: path
        required: true
    get:
      operationId: getReminderById
      summary: Busca um lembrete
      tags: 
        - Reminders
      responses:
        '201':
          description: 'ok'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'   
        '500':
          $ref: '#/components/responses/InternalServerError'
  
    put:
      operationId: updateReminder
      summary: Atualizar lembrete
      tags: 
        - Reminders
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderInput'
      responses:
        '204':
          description: 'Lembrete atualizado'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'   
        '500':
          $ref: '#/components/responses/InternalServerError'
  
    delete:
      operationId: removeReminder
      summary: Deleta lembrete
      tags: 
        - Reminders
      responses:
        '204':
          description: 'Lembrete removido'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderDetails'
        '404':
          $ref: '#/components/responses/NotFound'   
        '500':
          $ref: '#/components/responses/InternalServerError'

tags:
  - name: Reminders
    description: Lembretes

security:
  - ApiKey: []
  
servers:
  - description: 'Ambiente de produção'
    url: 'https://calendar.algaworks;com'
  - description: 'Ambiente local de desenvolvimento'
    url: 'http://localhost:8080'
    
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      name: X-Authorization
      in: header
      description: 'Token de autenticação'
  responses:
    NotFound:
      description: 'Recurso não encontrado'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
            
    BadRequest:
      description: 'Requisição inválida'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
            
    InternalServerError:
      description: 'Erro interno no servidor'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
            
  schemas:
    ReminderDetails:
      description: 'Lembrete detalhado'
      type: object
      required:
        - id
        - title
        - description
        - date
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 'Aula de Jiu Jistsu semanal'
        description:
          type: string
          example: 'Aula de Jiu Jitsu com 0 Flavio'
        date:
          type: string
          format: date
          example: '2022-12-30'
  
    ReminderSummary:
      description: 'Lembrete resumido'
      type: object
      required:
        - id
        - title
        - date
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 'Aula de Jiu Jistsu semanal'
        date:
          type: string
          format: date
          example: '2022-12-30'
  
    ReminderInput:
      description: 'Modelo para criação ou alteração de um lembrete'
      type: object
      required:
        - title
        - description
        - date
      properties:
        title:
          type: string
          example: 'Treino semanal'
        description:
          type: string
          example: 'Treino semanal turma da 17h'
        date:
          type: string
          format: date
          example: '2022-12-30'
  
    ErrorMessage:
        description: 'Representa um modelo de erro na requisção'
        type: object
        required:
          - status
          - timestamp
          - error
          - message
          - path
        properties:
          status:
            type: integer
            format: int32
            example: '400'
          timestamp:
            type: string
            format: date-time
            example: '2022-12-31-T18:09:027'
          error:
            type: string
            example: 'Dados inválidos'
          message:
            type: string
            example: 'Um ou mais campos inválidos'
          path:
            type: string
            example: '/path'
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
        
      